LayoutTests/canvas/philip/tests/2d.imageData.get.type-expected.txt

< [index] >
2d.imageData.get.type
getImageData() returns an ImageData object containing a CanvasPixelArray object
References: 2d.imageData.get.object
Actual output:
Passed
